move main to own file
